+++ apps/api/src/vendors/vendors.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { SequelizeModule } from '@nestjs/sequelize';
+import { Vendor } from './vendor.model';
+import { VendorsController } from './vendors.controller';
+import { VendorsService } from './vendors.service';
+
+@Module({
+  imports: [SequelizeModule.forFeature([Vendor])],
+  controllers: [VendorsController],
+  providers: [VendorsService],
+  exports: [VendorsService],
+})
+export class VendorsModule {}